+++ packages/ui-core/gulpfile.js
@@ -29,6 +29,7 @@ const indexTs = path.join(srcPath, 'index.ts');
 const baseLessSrc = path.join(srcPath, 'styles', 'base.less');
 const baseLessPackagePath = path.join(__dirname, 'styles');
 
+const doczReg = 'src/**/*.docz.{tsx,ts}';
 const testsReg = 'src/**/*.test.{tsx,ts}';
 const iconsReg = 'src/**/Icons.{tsx,ts}';
 
@@ -111,7 +112,7 @@ gulp.task('less:compile', () => {
                 .replace(/icons\//g, '')
                 .replace(/\//g, '-');
 
-            return `../icons/${newPath}`;
+            return `dist/icons/${newPath}`;
         }))
         .pipe(gulpLess(lessConfig))
         .pipe(dest(esPath))
@@ -126,8 +127,13 @@ gulp.task('less:copy-base', function() {
 gulp.task('less', gulp.series('less:compile', 'less:copy-base'));
 
 gulp.task('ts', () => {
-    const result = gulp.src(['src/**/*.{tsx,ts}', `!${testsReg}`, `!${iconsReg}`, './typings/*.d.ts'])
-        .pipe(ts(tsConfig));
+    const result = gulp.src([
+        'src/**/*.{tsx,ts}',
+        `!${doczReg}`,
+        `!${testsReg}`,
+        `!${iconsReg}`,
+        './typings/*.d.ts'
+    ]).pipe(ts(tsConfig));
 
     return merge(
         result.dts
@@ -144,7 +150,7 @@ gulp.task('ts', () => {
 });
 
 gulp.task('main', done => {
-    const components = glob.sync('src/components/**/*.{tsx,ts}', { ignore: testsReg });
+    const components = glob.sync('src/components/**/*.{tsx,ts}', { ignore: [testsReg, doczReg] });
     const utils = glob.sync('src/utils/*.ts', { ignore: testsReg });
     fs.writeFile(indexTs, generateIndex([...components, ...utils]), done);
 });
@@ -170,8 +176,9 @@ const generateIndex = files => {
     });
     const collator = new Intl.Collator();
     const sorted = components.sort((a, b) => collator.compare(a.name, b.name));
+    const getCnCreate = name => name === 'cn' ? `${name}, default as cnCreate` : name;
     const imports = sorted.map(({ name, path: cPath, ext: extension }) => {
-        return `export { default as ${name} } from '${cPath.replace(extension, '')}';`;
+        return `export { default as ${getCnCreate(name)} } from '${cPath.replace(extension, '')}';`;
     });
 
     return `${imports.join('\n')}`;