+++ packages/ui-core/src/components/BubbleHint/BubbleHint.test.tsx
@@ -17,9 +17,7 @@ describe('<BubbleHint />', () => {
 
     it('it renders BubbleHint', () => {
         jest.doMock('../../utils/detectTouch', () => {
-            return {
-                default: jest.fn(() => true),
-            };
+            return jest.fn(() => true);
         });
 
         // tslint:disable-next-line:no-shadowed-variable
@@ -71,7 +69,7 @@ describe('<BubbleHint />', () => {
         div.contains = () => { return true; };
 
         instance.getTrigger(div);
-        instance.handleClick({ target: {} } as React.SyntheticEvent<EventTarget>);
+        instance.handleClick({ target: {} } as React.MouseEvent<HTMLElement>);
         expect(onClick.mock.calls).toHaveLength(1);
     });
 
@@ -80,9 +78,8 @@ describe('<BubbleHint />', () => {
         const wrapper = shallow(<BubbleHint onClick={onClick} />);
         const instance = wrapper.instance() as BubbleHint;
 
-        // wrapper.instance().getTrigger({ contains() { return false; } });
         instance.getTrigger(div);
-        instance.handleClick({ target: {} } as React.SyntheticEvent<EventTarget>);
+        instance.handleClick({ target: {} } as React.MouseEvent<HTMLElement>);
         expect(onClick.mock.calls).toHaveLength(0);
     });