+++ packages/ui-core/src/components/Button/Button.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import * as PropTypes from 'prop-types';
-import cnCreate from 'utils/cn';
+import cnCreate from 'utils/cnCreate';
 import './Button.less';
 import Spinner from 'docIcons/spinner.svg';
 import detectTouch from 'utils/detectTouch';
@@ -52,7 +52,7 @@ export interface IButtonProps {
     showSpinner?: boolean;
     children?: JSX.Element[] | Element[] | JSX.Element | string | Element;
     /** Click event handler */
-    onClick?(e: React.SyntheticEvent<EventTarget>): void;
+    onClick?: (e: React.SyntheticEvent<EventTarget>) => void;
 }
 
 interface IButtonState {