+++ packages/ui-core/src/components/Carousel/Carousel.tsx
@@ -1,10 +1,11 @@
 import * as React from 'react';
 import * as PropTypes from 'prop-types';
-import cnCreate from 'utils/cn';
+import cnCreate from 'utils/cnCreate';
 import './Carousel.less';
 import CarouselArrow from './CarouselArrow';
 import Slider from 'react-slick';
-import throttle from 'lodash.throttle';
+
+const throttle = require('lodash.throttle');
 
 interface ICarouselOptions {
     slidesToShow: number;
@@ -22,6 +23,13 @@ interface ICarouselOptionsResponsive {
     settings: Pick<ICarouselOptions, 'slidesToShow' | 'arrows'>;
 }
 
+interface ICarouselClasses {
+    root?: string;
+    slider?: string;
+    leftArrow?: string;
+    rightArrow?: string;
+}
+
 export interface ICarouselProps {
     className?: string;
     options: ICarouselOptions;
@@ -29,11 +37,13 @@ export interface ICarouselProps {
     arrowColor?: string;
     /** Padding between slides */
     disablePaddingBetweenSlides?: boolean;
+    classes?: ICarouselClasses;
     children: any;
     onClickNext?: () => void;
     onClickPrev?: () => void;
     onAfterChange?: (index: number) => void;
     onBeforeChange?: (currentIndex: number, nextIndex: number) => void;
+    onSwipe?: (direction: string) => void;
 }
 
 interface ICarouselState {
@@ -60,6 +70,12 @@ class Carousel extends React.Component<ICarouselProps, ICarouselState> {
                 PropTypes.array,
             ])
         ),
+        classes: PropTypes.shape({
+            root: PropTypes.string,
+            slider: PropTypes.string,
+            leftArrow: PropTypes.string,
+            rightArrow: PropTypes.string,
+        }),
         theme: PropTypes.oneOf(['default', 'landing', 'showcase', 'lk']),
         arrowColor: PropTypes.oneOf(['white']),
         disablePaddingBetweenSlides: PropTypes.bool,
@@ -75,10 +91,10 @@ class Carousel extends React.Component<ICarouselProps, ICarouselState> {
         disablePaddingBetweenSlides: false,
     };
 
-    firstClientX: number;
-    clientX: number;
+    firstClientX: number = 0;
+    clientX: number = 0;
     noPassiveOption: any = { passive: false };
-    slider: any;
+    slider: any = null;
     throttledResizeEvents: () => void;
 
     constructor(props: ICarouselProps) {
@@ -244,24 +260,31 @@ class Carousel extends React.Component<ICarouselProps, ICarouselState> {
         );
     }
 
+    handleSwipe = (direction: string): void => {
+        const { onSwipe } = this.props;
+        onSwipe && onSwipe(direction);
+    }
+
     resizeEvents = (): void => {
         this.handleCarouselParams();
         this.checkIfSlidePositionWasLost();
     }
 
     renderArrows() {
-        const { theme, arrowColor } = this.props;
+        const { theme, arrowColor, classes = {} } = this.props;
         const { isPrevActive, isNextActive } = this.state;
+        const modPrevArrow = { 'arrow-prev': true, disabled: !isPrevActive, fill: arrowColor };
+        const modNextArrow = { 'arrow-next': true, disabled: !isNextActive, fill: arrowColor };
 
         return (
             <div className={cn('arrows', { theme: theme })}>
                 <CarouselArrow
-                    className={cn('arrow', { 'arrow-prev': true, disabled: !isPrevActive, fill: arrowColor })}
+                    className={cn('arrow', modPrevArrow, classes.leftArrow)}
                     onClick={this.handleClickPrev}
                     theme={theme}
                 />
                 <CarouselArrow
-                    className={cn('arrow', { 'arrow-next': true, disabled: !isNextActive, fill: arrowColor })}
+                    className={cn('arrow', modNextArrow, classes.rightArrow)}
                     onClick={this.handleClickNext}
                     theme={theme}
                 />
@@ -270,21 +293,23 @@ class Carousel extends React.Component<ICarouselProps, ICarouselState> {
     }
 
     render() {
-        const { className, options, theme, children, onBeforeChange, disablePaddingBetweenSlides } = this.props;
+        const { options, theme, children, onBeforeChange, disablePaddingBetweenSlides } = this.props;
         const { isArrows } = this.state;
-        const { arrows, ...carouselOptions } = options;
+        const { className = '', classes = {} } = this.props;
+        const modClasses = { theme, 'no-padding-between-slides': disablePaddingBetweenSlides };
+        const propsClassName = `${className} ${classes.root || ''}`.trim();
 
         return (
-            <div
-                className={cn('', { theme, 'no-padding-between-slides': disablePaddingBetweenSlides }, className)}
-            >
+            <div className={cn('', modClasses, propsClassName)}>
                 {isArrows && this.renderArrows()}
                 <Slider
-                    {...carouselOptions}
+                    {...options}
+                    className={classes.slider}
                     arrows={false}
                     ref={this.getSlider}
                     afterChange={this.handleChange}
                     beforeChange={onBeforeChange}
+                    onSwipe={this.handleSwipe}
                 >
                     {children}
                 </Slider>