+++ packages/ui-core/src/components/ContentArea/ContentArea.docz.tsx
@@ -1,0 +1,5 @@
+export const playgroundStyle = {
+    backgroundColor: '#E5F5EE',
+    paddingLeft: '0',
+    paddingRight: '0',
+};