+++ packages/ui-core/src/components/Grid/Grid.tsx
@@ -1,10 +1,10 @@
 import * as React from 'react';
 import * as PropTypes from 'prop-types';
 import './Grid.less';
-import cnCreate from 'utils/cn';
+import cnCreate from 'utils/cnCreate';
 import { IGridColumn } from './GridColumn';
 
-interface IProps {
+interface IGridProps {
     /** Alignment of all columns by horizontal axis */
     hAlign?: 'right' | 'center' | 'between' | 'around';
     /** Alignment of all columns by vertical axis */
@@ -19,7 +19,7 @@ interface IProps {
 }
 
 const cn = cnCreate('mfui-grid');
-class Grid extends React.Component<IProps, {}> {
+class Grid extends React.Component<IGridProps> {
     static propTypes = {
         hAlign: PropTypes.oneOf(['right', 'center', 'between', 'around']),
         vAlign: PropTypes.oneOf(['top', 'bottom', 'center', 'baseline']),
@@ -46,18 +46,18 @@ class Grid extends React.Component<IProps, {}> {
                         'gutters-left': guttersLeft,
                         'gutters-bottom': guttersBottom,
                     })}>
-                        {React.Children.map(children, (child: React.ReactElement<IGridColumn>) =>
-                            React.cloneElement(child, {
-                                className: cn(
-                                    'column',
-                                    {
-                                        'gutter-left': guttersLeft,
-                                        'gutter-bottom': guttersBottom,
-                                    },
-                                    child.props.className
-                                ),
-                            })
-                        )}
+                    {React.Children.map(children, (child) =>
+                        React.cloneElement(child, {
+                            className: cn(
+                                'column',
+                                {
+                                    'gutter-left': guttersLeft,
+                                    'gutter-bottom': guttersBottom,
+                                },
+                                child.props.className
+                            ),
+                        })
+                    )}
                 </div>
             </div>
         );