+++ packages/ui-core/src/components/TextLink/TextLink.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 import * as PropTypes from 'prop-types';
 import './TextLink.less';
 import Link from '../Link/Link';
-import cnCreate from 'utils/cn';
+import cnCreate from 'utils/cnCreate';
 
 interface ITextLinkProps {
     /** Color */
@@ -12,14 +12,14 @@ interface ITextLinkProps {
     /** Underline style */
     underlineStyle: 'solid' | 'dashed' | 'border' | 'none';
     /** Target - property tag <a> */
-    target: '_self' | '_blank' | '_parent' | '_top';
+    target?: '_self' | '_blank' | '_parent' | '_top';
     /** Link */
-    href: string;
+    href?: string;
     /** Custom class name */
     className?: string;
-    children: JSX.Element[] | Element[] | JSX.Element | string | Element;
+    children?: JSX.Element[] | Element[] | JSX.Element | string | Element;
     /** Click handler */
-    onClick?(e: React.SyntheticEvent<EventTarget>): void;
+    onClick?: (e: React.SyntheticEvent<EventTarget>) => void;
 }
 
 const cn = cnCreate('mfui-text-link');
@@ -51,14 +51,23 @@ class TextLink extends React.Component<Partial<ITextLinkProps>, {}> {
             underlineStyle,
             color,
             className,
-            ...props
+            target,
+            href,
+            onClick,
+            children,
         } = this.props;
 
         return (
-            <Link {...props} className={cn('', {
-                'underline-visibility': underlineVisibility,
-                'underline-style': underlineStyle, color,
-            }, className)} />
+            <Link
+                target={target}
+                href={href}
+                onClick={onClick}
+                children={children}
+                className={cn('', {
+                    'underline-visibility': underlineVisibility,
+                    'underline-style': underlineStyle, color,
+                }, className)}
+            />
         );
     }
 }