+++ packages/ui-core/src/components/Textarea/Textarea.docz.tsx
@@ -1,0 +1,5 @@
+import * as React from 'react';
+
+export const TextAreaWrapper: React.FC = ({ children }) => (
+    <div style={{ width: 250 }}>{children}</div>
+);