+++ packages/ui-shared/src/components/ProductTileRest/ProductTileRest.tsx
@@ -1,13 +1,10 @@
 import * as React from 'react';
 import * as PropTypes from 'prop-types';
-import { cn as cnCreate } from '@megafon/ui-core';
+import { cnCreate, Header, TextLink, Button, DropdownSocialList } from '@megafon/ui-core';
 import './ProductTileRest.less';
-import { Header, TextLink, Button, DropdownSocialList } from '@megafon/ui-core';
 
-type IButtonProps = Button['props'];
-
-const LinkTargetType = PropTypes.oneOf(['_self', '_blank', '_parent', '_top']);
 export type TLinkTargetType = '_self' | '_blank' | '_parent' | '_top';
+type IButtonProps = React.ComponentProps<typeof Button>;
 
 interface IProductTileRestProps {
     className?: string;
@@ -24,6 +21,7 @@ interface IProductTileRestProps {
     buyLinkTarget?: TLinkTargetType;
     buyButtonText: string;
     showBuyButton: boolean;
+
     /** button border */
     buttonBorder?: IButtonProps['border'];
     /** button font color */
@@ -46,6 +44,8 @@ interface IProductTileRestProps {
     onClickBuy: any;
     onClickMore: any;
 }
+
+const LinkTargetType = PropTypes.oneOf(['_self', '_blank', '_parent', '_top']);
 
 const cn = cnCreate('mfui-product-tile-rest');
 class ProductTileRest extends React.Component<IProductTileRestProps> {
@@ -183,7 +183,7 @@ class ProductTileRest extends React.Component<IProductTileRestProps> {
         const {
             payment: { title, value, unitExtra, unitValue, discount },
             packs,
-        } =  this.props;
+        } = this.props;
 
         return (
             <React.Fragment>
@@ -239,7 +239,7 @@ class ProductTileRest extends React.Component<IProductTileRestProps> {
         } = this.props;
 
         return (
-            <div className={cn('', {active: isActive}, className)}>
+            <div className={cn('', { active: isActive }, className)}>
                 <div className={cn('info')}>
                     <Header className={cn('header')} as="h3">{title}</Header>
                     {showMoreLink &&