+++ packages/ui-shared/CHANGELOG.md
@@ -2,6 +2,34 @@
 
 ## Unreleased
 
+## Release 1.0.30
+- **package.json**
+    - updated @megafon/ui-core package
+
+## Release 1.0.29
+- **package.json**
+    - updated @megafon/ui-core package
+
+## Release 1.0.28
+- **package.json**
+    - updated @megafon/ui-core package
+
+## Release 1.0.27
+- **package.json**
+    - updated @megafon/ui-core package
+
+## Release 1.0.26
+- **package.json**
+    - updated @megafon/ui-core package
+
+## Release 1.0.25
+- **package.json**
+    - updated @megafon/ui-core package
+
+## Release 1.0.24
+- **package.json**
+    - updated @megafon/ui-core package
+
 ## Release 1.0.23
 - **package.json**
     - updated @megafon/ui-core package